Testing

+++ c_lib/Stepper.h
@@ -39,13 +39,13 @@ void Initialize_Stepper( Stepper_t* p_step, uint16_t pos, int number_of_steps, i
  * @param p_step pointer to the desired stepper object to manipulate
  * @param Value speed value to update the stepper to (RPM)
  */
-void Stepper_Speed( Stepper_t* p_step, int Value);
+void Stepper_Speed( Stepper_t* p_step, float Value);
 
 /**
  * Function Stepper_Step takes an input of number of steps and moves the desired stepper motor
  * @param p_step pointer to the desired stepper object
  * @param steps desired number of steps
  */
-void Stepper_Step( Stepper_t* p_step, int steps );
+void Stepper_Step( Stepper_t* p_step, float steps );
 
 #endif
